pyproject.toml
[tool.poetry]
name = "diffusion"
version = "0.1.0"
description = ""
authors = ["KaiYamashita <kai-yama12345679@g.ecc.u-tokyo.ac.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.66.4"
torch = "^2.3.0"
deepspeed = "^0.14.2"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
python-dotenv = "^1.0.1"
wandb = "^0.17.0"
torchvision = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
